# tools/openapi.json
{
    "openapi":  "3.1.0",
    "info":  {
                 "title":  "Pappaliiga Stats API",
                 "description":  "REST API for CS2 tournament statistics",
                 "version":  "1.0.0"
             },
    "paths":  {
...
                  "/api/teams/{team_id}":  {
                                               "get":  {
                                                           "tags":  [
                                                                        "teams"
                                                                    ],
                                                           "summary":  "Get Team Info",
                                                           "description":  "Get basic team information.",
                                                           "operationId":  "get_team_info_api_teams__team_id__get",
                                                           "parameters":  [
                                                                              "@{name=team_id; in=path; required=True; schema=}"
                                                                          ],
                                                           "responses":  {
                                                                             "200":  "@{description=Successful Response; content=}",
                                                                             "422":  "@{description=Validation Error; content=}"
                                                                         }
                                                       }
                                           },
                  "/api/teams/{team_id}/seasons":  {
                                                       "get":  {
                                                                   "tags":  [
                                                                                "teams"
                                                                            ],
                                                                   "summary":  "Get Team Season Stats",
                                                                   "description":  "Get team stats across all seasons/divisions.",
                                                                   "operationId":  "get_team_season_stats_api_teams__team_id__seasons_get",
                                                                   "parameters":  [
                                                                                      "@{name=team_id; in=path; required=True; schema=}"
                                                                                  ],
                                                                   "responses":  {
                                                                                     "200":  "@{description=Successful Response; content=}",
                                                                                     "422":  "@{description=Validation Error; content=}"
                                                                                 }
                                                               }
                                                   },
                  "/api/teams/{team_id}/map-stats/{championship_id}":  {
                                                                           "get":  {
                                                                                       "tags":  [
                                                                                                    "teams"
                                                                                                ],
                                                                                       "summary":  "Get Team Map Stats",
                                                                                       "description":  "Get detailed per-map stats with deltas for a team in a specific championship.",
                                                                                       "operationId":  "get_team_map_stats_api_teams__team_id__map_stats__championship_id__get",
                                                                                       "parameters":  [
                                                                                                          "@{name=team_id; in=path; required=True; schema=}",
                                                                                                          "@{name=championship_id; in=path; required=True; schema=}"
                                                                                                      ],
                                                                                       "responses":  {
                                                                                                         "200":  "@{description=Successful Response; content=}",
                                                                                                         "422":  "@{description=Validation Error; content=}"
                                                                                                     }
                                                                                   }
                                                                       },
                  "/api/teams/":  {
                                      "get":  {
                                                  "tags":  [
                                                               "teams"
                                                           ],
                                                  "summary":  "List Teams",
                                                  "description":  "List all teams, optionally filtered by season/division.",
                                                  "operationId":  "list_teams_api_teams__get",
                                                  "parameters":  [
                                                                     "@{name=season; in=query; required=False; schema=; description=Filter by season}",
                                                                     "@{name=division; in=query; required=False; schema=; description=Filter by division}",
                                                                     "@{name=limit; in=query; required=False; schema=}"
                                                                 ],
                                                  "responses":  {
                                                                    "200":  "@{description=Successful Response; content=}",
                                                                    "422":  "@{description=Validation Error; content=}"
                                                                }
                                              }
                                  },
...
              },
...
}